+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-- Babel security fix
+# Additions
+
+- The cookies from the WS client's browser now can be accessed. (ZilaClient.cookies)
+- You can now set cookies from the serverside if the client is connected from a browser.
+- Checks for passing a function down to a client has been removed. The built-in JSON serializer ingores functions by default.